+++ src/components/Initiative/utils/getCategories.js
